requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
pydantic
requests
aiofiles
pymupdf